requirements.txt
aniso8601==7.0.0
asgiref==3.2.10
certifi==2020.6.20
cffi==1.14.5
chardet==3.0.4
contextvars==2.4
cryptography==3.4.7
defusedxml==0.6.0
diff-match-patch==20200713
Django==3.1.5
django-appconf==1.0.4
django-ckeditor==5.9.0
django-easy-select2==1.5.7
django-extensions==3.1.3
django-filter==2.4.0
django-graphql-jwt==0.3.1
django-graphql-social-auth==0.1.4
django-imagekit==4.0.2
django-import-export==2.3.0
django-js-asset==1.2.2
djangorestframework==3.12.4
et-xmlfile==1.0.1
geographiclib==1.52
geopy==2.2.0
googletrans==3.1.0a0
graphene==2.1.8
graphene-django==2.13.0
graphql-core==2.3.2
graphql-relay==2.0.1
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2021.8.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
immutables==0.15
jdcal==1.4.1
MarkupPy==1.14
oauthlib==3.1.1
odfpy==1.4.1
onesignal-python==0.1.0.dev1
onesignal-sdk==2.0.0
openpyxl==3.0.5
pilkit==2.0
Pillow==8.1.0
promise==2.3
pycparser==2.20
PyJWT==2.1.0
python-social-auth==0.3.6
python3-openid==3.2.0
pytz==2020.1
PyYAML==5.3.1
requests==2.12.1
requests-oauthlib==1.3.0
rfc3986==1.5.0
Rx==1.6.1
semantic-version==2.8.5
setuptools-rust==0.12.1
singledispatch==3.4.0.3
six==1.15.0
sniffio==1.2.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.3.1
tablib==2.0.0
toml==0.10.2
Unidecode==1.1.1
urllib3==1.25.10
whitenoise==5.3.0
xlrd==1.2.0
xlwt==1.3.0
gunicorn==20.1.0
dj-static==0.0.6
mysqlclient
django-extensions==3.1.3  # https://github.com/django-extensions/django-extensions
django-environ==0.6.0  # https://github.com/joke2k/django-environ
psycopg2==2.9.1
Werkzeug==2.0.1
dj-rest-auth==2.1.11 
djangorestframework-simplejwt==4.8.0
django-cors-headers==3.8.0
django-filter==2.4.0